+++ src/main/java/net/detalk/api/post/service/ProductPostService.java
@@ -18,8 +18,8 @@ import net.detalk.api.post.repository.ProductPostSnapshotAttachmentFileRepositor
 import net.detalk.api.post.repository.ProductPostSnapshotRepository;
 import net.detalk.api.post.repository.ProductPostSnapshotTagRepository;
 import net.detalk.api.product.domain.ProductLink;
-import net.detalk.api.domain.exception.DuplicateCreatePostException;
-import net.detalk.api.domain.exception.InvalidPageSizeException;
+import net.detalk.api.post.domain.exception.DuplicateCreatePostException;
+import net.detalk.api.support.error.InvalidPageSizeException;
 import net.detalk.api.post.domain.exception.InvalidRecommendCountRequest;
 import net.detalk.api.post.domain.exception.ProductPostForbiddenException;
 import net.detalk.api.post.domain.exception.ProductPostNotFoundException;
